+++ R/metabin.R
@@ -107,6 +107,9 @@
 #'   (see \code{\link{meta-package}}).
 #' @param level.ma The level used to calculate confidence intervals
 #'   for meta-analysis estimates.
+#' @param method.common.ci A character string indicating which method
+#'   is used to calculate confidence interval and test statistic for
+#'   common effect estimate (see \code{\link{meta-package}}).
 #' @param method.random.ci A character string indicating which method
 #'   is used to calculate confidence interval and test statistic for
 #'   random effects estimate (see \code{\link{meta-package}}).
@@ -675,6 +678,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                     method.I2 = gs("method.I2"),
                     #
                     level.ma = gs("level.ma"),
+                    method.common.ci = gs("method.common.ci"),
                     method.random.ci = gs("method.random.ci"),
                     adhoc.hakn.ci = gs("adhoc.hakn.ci"),
                     ##
@@ -765,6 +769,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
   missing.level.ma <- missing(level.ma)
   missing.common <- missing(common)
   missing.random <- missing(random)
+  missing.method.common.ci <- missing(method.common.ci)
   missing.method.random.ci <- missing(method.random.ci)
   missing.method.I2 <- missing(method.I2)
   #
@@ -794,6 +799,16 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
   #
   is.glmm <- method == "GLMM"
   is.lrp <- method == "LRP"
+  #
+  method.common.ci <- setchar(method.common.ci, gs("meth4common.ci"))
+  #
+  if (method != "Inverse" & method.common.ci == "IVhet") {
+    if (!missing.method.common.ci)
+      warning("Argument 'method.common.ci = \"IVhet\"' only available ",
+              "if 'method = \"Inverse\".",
+              call. = FALSE)
+    method.common.ci <- "classic"
+  }
   #
   if (missing.method.tau) {
     if (is.lrp)
@@ -1987,6 +2002,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                method.I2 = method.I2,
                #
                level.ma = level.ma,
+               method.common.ci = method.common.ci,
                method.random.ci = method.random.ci,
                adhoc.hakn.ci = adhoc.hakn.ci,
                ##