+++ R/metainf.R
@@ -357,6 +357,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                    tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                    ##
                    level.ma = x$level.ma,
+                   method.common.ci = x$method.common.ci,
                    method.random.ci = x$method.random.ci,
                    adhoc.hakn.ci = x$adhoc.hakn.ci,
                    #
@@ -382,6 +383,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                     tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                     ##
                     level.ma = x$level.ma,
+                    method.common.ci = x$method.common.ci,
                     method.random.ci = x$method.random.ci,
                     adhoc.hakn.ci = x$adhoc.hakn.ci,
                     #
@@ -405,6 +407,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                    tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                    ##
                    level.ma = x$level.ma,
+                   method.common.ci = x$method.common.ci,
                    method.random.ci = x$method.random.ci,
                    adhoc.hakn.ci = x$adhoc.hakn.ci,
                    #
@@ -427,6 +430,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                    tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                    ##
                    level.ma = x$level.ma,
+                   method.common.ci = x$method.common.ci,
                    method.random.ci = x$method.random.ci,
                    adhoc.hakn.ci = x$adhoc.hakn.ci,
                    #
@@ -455,6 +459,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                    tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                    ##
                    level.ma = x$level.ma,
+                   method.common.ci = x$method.common.ci,
                    method.random.ci = x$method.random.ci,
                    adhoc.hakn.ci = x$adhoc.hakn.ci,
                    #
@@ -478,6 +483,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                     tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                     ##
                     level.ma = x$level.ma,
+                    method.common.ci = x$method.common.ci,
                     method.random.ci = x$method.random.ci,
                     adhoc.hakn.ci = x$adhoc.hakn.ci,
                     #
@@ -504,6 +510,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                     tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                     ##
                     level.ma = x$level.ma,
+                    method.common.ci = x$method.common.ci,
                     method.random.ci = x$method.random.ci,
                     adhoc.hakn.ci = x$adhoc.hakn.ci,
                     #
@@ -529,6 +536,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
                     tau.preset = x$tau.preset, TE.tau = x$TE.tau,
                     ##
                     level.ma = x$level.ma,
+                    method.common.ci = x$method.common.ci,
                     method.random.ci = x$method.random.ci,
                     adhoc.hakn.ci = x$adhoc.hakn.ci,
                     #
@@ -841,6 +849,7 @@ metainf.meta <- function(x, pooled, sortvar, prediction, overall = x$overall,
               method = x$method,
               method.random = x$method.random,
               #
+              method.common.ci = x$method.common.ci,
               method.random.ci = x$method.random.ci,
               adhoc.hakn.ci = x$adhoc.hakn.ci,
               #