+++ R/update.meta.R
@@ -75,6 +75,9 @@
 #'   null hypothesis.
 #' @param level.ma The level used to calculate confidence intervals
 #'   for meta-analysis estimates.
+#' @param method.common.ci A character string indicating which method
+#'   is used to calculate confidence interval and test statistic for
+#'   common effect estimate (see \code{\link{meta-package}}).
 #' @param method.random.ci A character string indicating which method
 #'   is used to calculate confidence interval and test statistic for
 #'   random effects estimate (see \code{\link{meta-package}}).
@@ -324,6 +327,7 @@ update.meta <- function(object,
                         random = object$random,
                         overall = object$overall,
                         overall.hetstat = object$overall.hetstat,
+                        method.common.ci = object$method.common.ci,
                         method.random.ci = object$method.random.ci,
                         adhoc.hakn.ci = object$adhoc.hakn.ci,
                         method.predict = object$method.predict,
@@ -440,6 +444,7 @@ update.meta <- function(object,
   metaprop <- inherits(object, "metaprop")
   metarate <- inherits(object, "metarate")
   ##
+  missing.method.common.ci <- missing(method.common.ci)
   missing.method.random.ci <- missing(method.random.ci)
   missing.adhoc.hakn.ci <- missing(adhoc.hakn.ci)
   missing.text.random <- missing(text.random)
@@ -799,6 +804,13 @@ update.meta <- function(object,
       }
     }
   }
+  #
+  if (update_needed(object$version, 8, 2, verbose)) {
+    ##
+    ## Changes for meta objects with version < 8.2
+    ##
+    object$method.common.ci <- "classic"
+  }
   
   
   ##
@@ -848,7 +860,9 @@ update.meta <- function(object,
   random <- deprecated(random, missing.random, args, "comb.random",
                        warn.deprecated)
   chklogical(random)
-  ##
+  #
+  method.common.ci <- setchar(method.common.ci, gs("meth4common.ci"))
+  #
   method.random.ci <-
     deprecated(method.random.ci, missing.method.random.ci,
                args, "hakn", warn.deprecated)
@@ -937,6 +951,7 @@ update.meta <- function(object,
                     type = type, n.iter.max = n.iter.max,
                     level = level, level.ma = level.ma,
                     common = common, random = random,
+                    method.common.ci = method.common.ci,
                     method.random.ci = method.random.ci,
                     adhoc.hakn.ci = adhoc.hakn.ci,
                     method.tau = method.tau, method.tau.ci = method.tau.ci,
@@ -1137,8 +1152,13 @@ update.meta <- function(object,
   ##
   if (!is.null(subgroup.name))
     chkchar(subgroup.name, length = 1)
+  #
+  # Check arguments for common effect model
+  #
+  if (!missing.method.common.ci & missing.common)
+    common <- TRUE
   ##
-  ## Check variables for random effects model(s)
+  ## Check arguments for random effects model(s)
   ##
   if (!(missing.method.random.ci & missing.adhoc.hakn.ci) & missing.random)
     random <- TRUE
@@ -1311,6 +1331,7 @@ update.meta <- function(object,
                  common = common, random = random,
                  overall = overall, overall.hetstat = overall.hetstat,
                  ##
+                 method.common.ci = method.common.ci,
                  method.random.ci = method.random.ci,
                  adhoc.hakn.ci = adhoc.hakn.ci,
                  method.predict = method.predict,
@@ -1429,6 +1450,7 @@ update.meta <- function(object,
                   common = common, random = random,
                   overall = overall, overall.hetstat = overall.hetstat,
                   ##
+                  method.common.ci = method.common.ci,
                   method.random.ci = method.random.ci,
                   adhoc.hakn.ci = adhoc.hakn.ci,
                   method.predict = method.predict,
@@ -1484,6 +1506,7 @@ update.meta <- function(object,
                  common = common, random = random,
                  overall = overall, overall.hetstat = overall.hetstat,
                  ##
+                 method.common.ci = method.common.ci,
                  method.random.ci = method.random.ci,
                  adhoc.hakn.ci = adhoc.hakn.ci,
                  method.predict = method.predict,
@@ -1566,6 +1589,7 @@ update.meta <- function(object,
                  overall = overall, overall.hetstat = overall.hetstat,
                  ##
                  level.ma = level.ma,
+                 method.common.ci = method.common.ci,
                  method.random.ci = method.random.ci,
                  adhoc.hakn.ci = adhoc.hakn.ci,
                  method.predict = method.predict,
@@ -1700,6 +1724,7 @@ update.meta <- function(object,
                  common = common, random = random,
                  overall = overall, overall.hetstat = overall.hetstat,
                  ##
+                 method.common.ci = method.common.ci,
                  method.random.ci = method.random.ci,
                  adhoc.hakn.ci = adhoc.hakn.ci,
                  method.predict = method.predict,
@@ -1777,6 +1802,7 @@ update.meta <- function(object,
                   common = common, random = random,
                   overall = overall, overall.hetstat = overall.hetstat,
                   ##
+                  method.common.ci = method.common.ci,
                   method.random.ci = method.random.ci,
                   adhoc.hakn.ci = adhoc.hakn.ci,
                   method.predict = method.predict,
@@ -1854,6 +1880,7 @@ update.meta <- function(object,
                   common = common, random = random,
                   overall = overall, overall.hetstat = overall.hetstat,
                   ##
+                  method.common.ci = method.common.ci,
                   method.random.ci = method.random.ci,
                   adhoc.hakn.ci = adhoc.hakn.ci,
                   method.predict = method.predict,
@@ -1934,6 +1961,7 @@ update.meta <- function(object,
                   common = common, random = random,
                   overall = overall, overall.hetstat = overall.hetstat,
                   ##
+                  method.common.ci = method.common.ci,
                   method.random.ci = method.random.ci,
                   adhoc.hakn.ci = adhoc.hakn.ci,
                   method.predict = method.predict,